+++ src/App.tsx
@@ -4,13 +4,12 @@ import { UserWarning } from './UserWarning';
 import { USER_ID, getTodos } from './api/todos';
 import { Todo } from './types/Todo';
 
-import { TodoList } from './components/TodoList/TodoList.';
-import { Header } from './components/Header/Header';
-import { TodoFilter } from './components/TodoFilter/TodoFilter';
-
 import { FilterTodos } from './types/FilterTodos';
 import { Errors } from './types/Errors';
 import { ErrorNotification } from './components/ErrorNotification';
+import { TodoFilter } from './components/TodoFilter';
+import { Header } from './components/Header';
+import { TodoList } from './components/TodoList';
 
 function handleFilteredTodos(todos: Todo[], filterSelected: string) {
   const filteredTodos = [...todos];
